% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSimulation.R
\encoding{UTF-8}
\name{addSimulation}
\alias{addSimulation}
\title{function to generate simulated communities to add on ongoing restoration}
\usage{
addSimulation(
  trait,
  ava,
  und,
  it,
  rich,
  cwm,
  cwv,
  rao,
  cost,
  dens,
  stan,
  ref,
  rest,
  max_add,
  min_p,
  phi = 1
)
}
\arguments{
\item{trait}{data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{vector indicating availability of species (1 or 0)}

\item{und}{vector indicating undesired species (1 or 0)}

\item{it}{number of iterations (communities) for each area ongoing restoration}

\item{rich}{range of richness values}

\item{cwm}{vector with traits to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{rao}{vector with traits to calculate Rao Quadratic Entropy, or distance matrix (class dist)}

\item{cost}{vector of species cost per individual}

\item{dens}{vector of planting density}

\item{stan}{which parameters should be standardized by the maximum? Must provide trait names}

\item{ref}{matrix with proportion of species in the reference sites. NAs not accepted.}

\item{rest}{matrix with proportion of species in the restored sites. NAs not accepted.}

\item{max_add}{maximum proportion that can be added to restored areas.}

\item{min_p}{minimum species proportion to consider that species occurs in the area}
}
\description{
generates simulated communities and calculate its parameters.
}
