% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjString.R
\encoding{UTF-8}
\name{adjString}
\alias{adjString}
\title{Internal function to adjust a vector of logical tests to be evaluate on a data.frame}
\usage{
adjString(prefix, logicTest)
}
\arguments{
\item{prefix}{The name of data.frame object.}

\item{logicTest}{The vector of logical tests.}
}
\value{
The vector of logical tests adjusted.
}
\description{
Internal function to adjust a vector of logical tests to be evaluate on a data.frame
}
\seealso{
\code{\link{selectCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
