% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/appServer.R, R/appUI.R
\docType{data}
\encoding{UTF-8}
\name{app}
\alias{app}
\alias{appServer}
\alias{appUI}
\title{app}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
appServer(input, output, session)

appUI
}
\arguments{
\item{input}{Input list for the UI.}

\item{output}{Output list for the UI.}

\item{session}{Environment for the UI.}
}
\description{
app
}
\keyword{datasets}
