% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCWV.R
\encoding{UTF-8}
\name{calcCWV}
\alias{calcCWV}
\title{Internal function to calculate Community Weighted Variance (CWV)}
\usage{
calcCWV(x, traitSub)
}
\arguments{
\item{x}{A matrix with species proportions in the reference sites. NAs not accepted.}

\item{traitSub}{Data frame or matrix with species traits. Traits as columns and species as rows.}
}
\value{
A matrix with Community Weighted Variance.
}
\description{
The Community Weighted Variance (CWV) is based on the function matrix.t from the package SYNCSA.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
