% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMF.R
\encoding{UTF-8}
\name{calcMF}
\alias{calcMF}
\title{Internal function to calculate indices on multifunctionality matrix}
\usage{
calcMF(x)
}
\arguments{
\item{x}{A data frame with binary multifunctionality tests.}
}
\value{
A list with the elements:
\item{landscape}{Landscape multifunctionality.}
\item{beta}{Rao's quadratic entropy calculated on multifunctionality matrix.}
\item{average}{Average number of functions per site.}
}
\description{
Calculate the landscape and beta multifunctionality indices and average number of functions per site.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
