% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPar.R
\encoding{UTF-8}
\name{calcPar}
\alias{calcPar}
\title{calcPar}
\usage{
calcPar(
  x,
  trait,
  ava,
  cwm,
  cwv,
  rao,
  cost,
  dens,
  stan,
  ref = NULL,
  supplementary = NULL
)
}
\description{
calcPar
}
