% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRAO.R
\encoding{UTF-8}
\name{calcRAO}
\alias{calcRAO}
\title{Internal function to calculate Rao's quadratic entropy (RAO)}
\usage{
calcRAO(comm, dis = NULL)
}
\arguments{
\item{comm}{A matrix with species proportions in the reference sites. NAs not accepted.}

\item{dis}{Distances among species.}
}
\value{
A matrix with among-community diversities excluding within-community diversity.
}
\description{
Rao's quadratic entropy (RAO) is based on the function discomQE from the package adiv.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
