% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDissimilarity.R
\encoding{UTF-8}
\name{calculateDissimilarity}
\alias{calculateDissimilarity}
\title{function to calculate functional dissimilarity (distCalculation)}
\usage{
calculateDissimilarity(x, trait)
}
\arguments{
\item{trait}{data frame or matrix with species traits. Traits as columns and species as rows.}

\item{sim}{data frame or matrix with species proportions in simulated communities. Species as columns and simulated communities as rows.}

\item{dis}{matrix of trait distances}

\item{ref}{data frame or matrix with species proportions in reference sites. Species as columns and simulated communities as rows.}
}
\description{
Calculate functional dissimilarity between simulated communities and reference sites
}
