% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMultifunctionality.R
\encoding{UTF-8}
\name{calculateMultifunctionality}
\alias{calculateMultifunctionality}
\title{function to calculate multifunctionality (multCalculation)}
\usage{
calculateMultifunctionality(x, tests, where = "global")
}
\arguments{
\item{x}{x}

\item{tests}{tests}

\item{where}{where}
}
\description{
function to calculate multifunctionality
}
