% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkReference.R, R/print.simRestCheck.R
\encoding{UTF-8}
\name{checkReference}
\alias{checkReference}
\alias{print.simRestCheck}
\title{checkReference}
\usage{
checkReference(
  reference,
  trait,
  cwm,
  cwv,
  rao,
  stan,
  supplementary = NULL,
  props = NULL
)

\method{print}{simRestCheck}(x, ...)
}
\arguments{
\item{reference}{A matrix with species proportions in the reference sites. NAs not accepted.}

\item{trait}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{cwm}{A vector with traits names to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{cwv}{A vector with traits names to calculate Community Weighted Variance (CWV). One CWV is calculated for each trait.}

\item{rao}{A vector with traits names to calculate Rao Quadratic Entropy, or distance matrix (class dist).}

\item{stan}{A vector with parameters names to specify which parameters should be standardized by the maximum.}

\item{supplementary}{A matrix with species proportions in the supplementary sites. NAs not accepted. (default supplementary = NULL).}

\item{props}{Numeric vector of probabilities with values in between 0 and 1 to produces sample quantiles corresponding to the given probabilities (default props = NULL).}
}
\description{
checkReference
}
\keyword{MainFunction}
