% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkResbiotaData.R
\encoding{UTF-8}
\name{checkResbiotaData}
\alias{checkResbiotaData}
\title{Internal function to check data to resbiota package}
\usage{
checkResbiotaData(
  traits = NULL,
  restComp = NULL,
  restGroup = NULL,
  reference = NULL,
  supplementary = NULL,
  traitsDist = NULL,
  cooccur = NULL,
  asList = TRUE
)
}
\arguments{
\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{restComp}{A matrix with species proportions in the restoration sites. NAs not accepted.}

\item{restGroup}{Data frame or matrix with complementary information for restoration sites.}

\item{reference}{A matrix with species proportions in the reference sites.}

\item{supplementary}{A matrix with species proportions in the supplementary sites.}

\item{traitsDist}{A distance matrix between species, based on functional trait values.}

\item{asList}{A logical argument to return the results in as list format.}
}
\value{
A list with the elements:
\item{checkStatus}{The global status with the type of alert.}
\item{checkMessage}{A vector with error messages.}
\item{checkWarning}{A vector with warning messages.}
}
\description{
Internal function to check data to resbiota package
}
\seealso{
\code{\link{selectCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
