% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classSimplerVector.R
\encoding{UTF-8}
\name{classSimplerVector}
\alias{classSimplerVector}
\title{Internal function to return the vector class.}
\usage{
classSimplerVector(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The vector class.
}
\description{
Internal function to return the vector class.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
