% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectNames.R
\encoding{UTF-8}
\name{collectNames}
\alias{collectNames}
\title{Internal function to collect the names of lists}
\usage{
collectNames(x, prefix = NULL)
}
\arguments{
\item{x}{A list object from which to extract element names.}

\item{prefix}{A prefix to include in collected names (default prefix = NULL).}
}
\value{
A character vector with collected names.
}
\description{
Internal function to collect the names of lists
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{InternalFunction}
