% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeParameters.R, R/computeDissimilarity.R,
%   R/computeMultifunctionality.R
\encoding{UTF-8}
\name{computeParameters}
\alias{computeParameters}
\alias{computeDissimilarity}
\alias{computeMultifunctionality}
\title{Compute functional parameters in communities}
\usage{
computeParameters(
  x,
  trait,
  ava,
  cwm,
  cwv,
  rao,
  cost,
  dens,
  stan,
  reference = NULL,
  supplementary = NULL
)

computeDissimilarity(x, trait)

computeMultifunctionality(x, tests)
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to perform calculate communities parameters.}

\item{trait}{data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{A vector indicating trait name which indicates the availability of species (1 or 0) in trait data.}

\item{cwm}{A vector with traits names to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{cwv}{A vector with traits names to calculate Community Weighted Variance (CWV). One CWV is calculated for each trait.}

\item{rao}{A vector with traits names to calculate Rao Quadratic Entropy, or distance matrix (class dist).}

\item{cost}{A vector with trait name with of species cost per individual}

\item{dens}{A vector with trait name with species planting density}

\item{stan}{A vector with parameters names to specify which parameters should be standardized by the maximum.}

\item{reference}{A matrix with species proportions in the reference sites. NAs not accepted. (default reference = NULL)}

\item{supplementary}{A matrix with species proportions in the supplementary sites. NAs not accepted. (default supplementary = NULL).}

\item{tests}{A vector with multifunctionality criteria to be performed.}
}
\value{
A list (class "simRest" or "simRestSelect") with the elements:
}
\description{
Calculate basic parameters in each community: richness, count species unavailable, Community Weighted Mean, Community Weighted Variance, Rao Quadratic Entropy, functional dissimilarity and multifunctionality.
}
