% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comSelection.R
\encoding{UTF-8}
\name{comSelection}
\alias{comSelection}
\title{function to select communities}
\usage{
comSelection(param, comp, tests)
}
\arguments{
\item{param}{data.frame with parameters of simulated communities}

\item{comp}{data.frame with compositions of simulated communities}

\item{tests}{list with tests to be performed}
}
\description{
select simulated communities based on tests provided and add information to x
}
