% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comSimulation2.R
\encoding{UTF-8}
\name{comSimulation2}
\alias{comSimulation2}
\title{function to generate simulated communities}
\usage{
comSimulation2(
  trait,
  ava,
  it,
  rich,
  cwm,
  cwv,
  rao,
  cost,
  dens,
  ref,
  min_p,
  phi = 1
)
}
\arguments{
\item{trait}{data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{vector indicating availability of species (1 or 0)}

\item{it}{number of iterations (communities)}

\item{rich}{range of richness values}

\item{cwm}{vector with traits to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{cwv}{vector vector with trait names to calculate Community Weighted Variance. One CWV is calculated for each trait.}

\item{rao}{vector with traits to calculate Rao Quadratic Entropy, or distance matrix (class dist)}

\item{cost}{vector of species cost per individual}

\item{dens}{vector of planting density}

\item{ref}{matrix with proportion of species in the reference sites. NAs not accepted.}

\item{min_p}{min_p}

\item{phi}{phi}
}
\description{
generates simulated communities and calculate its parameters.
}
