% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\encoding{UTF-8}
\name{createReport}
\alias{createReport}
\title{Creates a data profiling report}
\usage{
createReport(x, props = NULL)
}
\arguments{
\item{x}{A data.frame or matrix with the input data.}

\item{props}{Numeric vector of probabilities with values in between 0 and 1 to produces sample quantiles corresponding to the given probabilities (default props = NULL).}
}
\value{
The report in html format.
}
\description{
Generate a data profiling report with descriptive statistics and univariate graphs.
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{computeParameters}}, \code{\link{selectCommunities}},
\code{\link{extractResults}}, \code{\link{viewResults}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{MainFunction}
