% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\encoding{UTF-8}
\name{extractResults}
\alias{extractResults}
\title{extractResults}
\usage{
extractResults(x, trait, ava, type = "unavailableSpecies")
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to visualize results}

\item{trait}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{A vector indicating trait name which indicates the availability of species (1 or 0) in trait data.}

\item{type}{unavailableSpecies}
}
\description{
Check the unavailable species that were present in selected communities these are the species necessary to achieve the thresholds when selected communities have at least one species not available on the market
}
\keyword{MainFunction}
