% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCWV.R
\encoding{UTF-8}
\name{FCWV}
\alias{FCWV}
\title{Internal function to calculate Community Weighted Variance (CWV)}
\usage{
FCWV(x, traitSub)
}
\arguments{
\item{x}{A matrix with species proportions in the reference sites. NAs not accepted.}

\item{traitSub}{Data frame or matrix with species traits. Traits as columns and species as rows.}
}
\value{
A matrix with Community Weighted Variance.
}
\description{
Internal function to calculate Community Weighted Variance (CWV)
}
\keyword{Auxiliary}
