% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCombinations.R
\encoding{UTF-8}
\name{makeCombinations}
\alias{makeCombinations}
\title{Internal function to make groups combinations}
\usage{
makeCombinations(x, groups = NULL, minSubset = 1, maxSubset = 1)
}
\arguments{
\item{x}{A vector with elements to make the combinations.}

\item{groups}{A vector with the groups which the element belongs.}

\item{minSubset}{Minimum elements in each subset.}

\item{maxSubset}{Maximum elements in each subset.}
}
\value{
A binary matrix with all elements combinations.
}
\description{
Make groups combinations with groups restriction.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
