% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multCalculation.R
\encoding{UTF-8}
\name{multCalculation}
\alias{multCalculation}
\title{function to calculate multifunctionality}
\usage{
multCalculation(x, th, bel)
}
\arguments{
\item{x}{list of table with parameters of different areas}

\item{th}{named vector of thresholds. Names must match x columns}

\item{bel}{character vector indicating functions that must be below threshold. Must match x columns.}
}
\description{
function to calculate multifunctionality
}
