% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiseSelection.R
\encoding{UTF-8}
\name{optimiseSelection}
\alias{optimiseSelection}
\title{Optimise community selection based on a multi-site approach}
\usage{
optimiseSelection(
  x,
  siteGroup = NULL,
  includeReference = TRUE,
  maxComb = 1000,
  calcTaxonomicBeta = TRUE,
  method = "betaRao",
  traits = NULL,
  beta = NULL
)
}
\arguments{
\item{x}{A object of class "simRestSelect" to perform the optimisation.}

\item{siteGroup}{Character vector specifying a parameter name to define simulation site groups.}

\item{includeReference}{Logical argument specifying whether to include reference sites in index calculations (default includeReference = TRUE).}

\item{maxComb}{Maximum number of simulation combinations to generate (default maxComb = 1000). If the total unique combinations exceed this limit, a random sample of combinations is generated.}

\item{calcTaxonomicBeta}{Logical argument to specify if calculates taxonomic beta diversity (default calcTaxonomicBeta = TRUE).}

\item{method}{Method to calculate the beta diversity, partial match to "betaRao", "betaJaccard" or "betaSoerensen" (default method = "betaRao") .}

\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{beta}{Character vector specifying trait names to calculate beta diversity, or distance matrix (class dist). This argument can be a list to calculate multiple beta indices using different trait sets or species distance matrices.}
}
\value{
A list (class "simRestSelect") with the elements:
\item{call}{The arguments used.}
\item{selection$composition}{A matrix with species composition for selected communities.}
\item{selection$group}{A data frame with complementary information for selected sites.}
\item{selection$baseline}{A matrix with with baseline species composition for simulated communities (contains all zeros when restComp is not provided.)}
\item{selection$results}{A data frame with calculated parameters in each selected community.}
\item{selection$multifunctionality}{A data frame with binary multifunctionality tests.}
\item{selection$thresholds}{A vector with the count of selected communities at each threshold. When simulations are merged, it is not shown.}
\item{selection$multisite$combinations}{A binary matrix with communities combinations.}
\item{selection$multisite$results}{A data frame with calculated parameters in each communities combinations.}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{reference$multifunctionality}{A data frame with binary multifunctionality tests to reference sites.}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
\item{supplementary$multifunctionality}{A data frame with binary multifunctionality tests to supplementary sites.}
}
\description{
Calculate the functional parameters of selected simulated community sets to optimise multi-site simulation selections. This function generates all unique combinations of one simulated community per restoration site and calculates multi-site indices for each combination.
}
\examples{
data("cerrado")
head(cerrado$traits)
# Simulation
scenario <- simulateCommunities(traits = cerrado$traits,
                                restComp = cerrado$restoration,
                                maxDiver = c("SLA", "Height", "Seed"),
                                constCWM = "BT",
                                rich = c(10, 15),
                                it = 100)
scenario
# Compute functional parameters
scenario <- computeParameters(x = scenario,
                              traits = cerrado$traits,
                              cwm = "BT",
                              rao = c("SLA", "Height", "Seed"),
                              reference = cerrado$reference)
scenario
# Compute multifunctionality
scenario <- computeMultifunctionality(scenario,
                                      tests = c("CWM_BT > 5.9",
                                                "rao > 0.2"))
scenario
# Select communities - Priority selection
scenarioSelected <- selectCommunities(x = scenario,
                                      testsPriority = c("simpson > 0.92",
                                                        "CWM_BT > 6"),
                                      siteGroup = "Site",
                                      singleSelection = FALSE)
scenarioSelected
# Optimise selection
scenarioSelectedMultisite <- optimiseSelection(scenarioSelected,
                                               siteGroup = "Site",
                                               traits = cerrado$traits,
                                               beta = c("SLA", "Height", "Seed"),
                                               calcTaxonomicBeta = TRUE)
scenarioSelectedMultisite
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{computeParameters}}, \code{\link{selectCommunities}},
\code{\link{extractResults}}, \code{\link{viewResults}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{MainFunction}
