% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randDivision.R
\encoding{UTF-8}
\name{randDivision}
\alias{randDivision}
\title{Internal function to randomly divide a total number into groups}
\usage{
randDivision(n, k)
}
\arguments{
\item{n}{A integer to divide.}

\item{k}{Number of groups.}
}
\value{
A numeric vector with the random division.
}
\description{
Internal function to randomly divide a total number into groups
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{propMatrix}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
