% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readjustAbundance.R
\encoding{UTF-8}
\name{readjustAbundance}
\alias{readjustAbundance}
\title{Internal function to randomly adjust vectors of species abundance or proportions}
\usage{
readjustAbundance(abund0, minAbund, method)
}
\arguments{
\item{abund0}{Numeric vector of species abundance or proportions to be adjusted.}

\item{minAbund}{Minimum abundance or proportion threshold to maintain for each species in simulated communities.}

\item{method}{The method used to obtain the samples, "proportions" or "individuals".}
}
\value{
A vector of adjusted species abundance or proportions.
}
\description{
Internal function to randomly adjust vectors of species abundance or proportions
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{generateCommunityMatrices}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
