% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizeMatrix.R, R/makeMatrixTemplate.R
\encoding{UTF-8}
\name{reorganizeMatrix}
\alias{reorganizeMatrix}
\alias{makeMatrixTemplate}
\title{reorganizeMatrix}
\usage{
reorganizeMatrix(template, x, fillNA = FALSE)

makeMatrixTemplate(...)
}
\arguments{
\item{template}{}

\item{x}{}

\item{fillNA}{}
}
\description{
reorganizeMatrix
}
\keyword{Auxiliary}
