% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resbiota-package.R
\docType{package}
\encoding{UTF-8}
\name{resbiota-package}
\alias{resbiota}
\alias{resbiota-package}
\title{resbiota - Restore Ecosystem Services and Biodiversity through Trait-based Approaches}
\description{
Create restoration solutions to assemble communities from a regional species pool to achieve multiple functional targets based on simulation.
}
\details{
The function in the resbiota package contains an integrative set of tools to create restoration solutions to assemble communities from a regional 
species pool. The framework allows achieving functional targets following multiple approaches from empty communities and/or adding species to ongoing 
restoration sites. Details are presented in the main function \code{\link{simulateCommunities}} and the vignettes can be browsed with `browseVignettes("resbiota")`.
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\author{
INCLUIR
}
