% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resbiotaRun.R
\encoding{UTF-8}
\name{resbiotaRun}
\alias{resbiotaRun}
\title{Launch the resbiota Shiny Application}
\usage{
resbiotaRun(launchBrowser = TRUE, ...)
}
\arguments{
\item{launchBrowser}{Shiny application launch mode. See runApp function (default launchBrowser = TRUE).}

\item{...}{Arguments used in the runApp function.}
}
\description{
Launches a local version of the resbiota application.
}
