% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resSummary.R
\encoding{UTF-8}
\name{resSummary}
\alias{resSummary}
\title{Internal function to calculate descriptive statistics}
\usage{
resSummary(x, props = NULL, ...)
}
\arguments{
\item{x}{A numeric vector to calculate descriptive statistics.}

\item{props}{Numeric vector of probabilities with values in between 0 and 1 to produces sample quantiles corresponding to the given probabilities (default props = NULL).}
}
\value{
A data frame with descriptive statistics: minimum, mean, median, maximum and quantiles
}
\description{
Internal function to calculate descriptive statistics
}
\keyword{Auxiliary}
