% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restauraR-package.R
\docType{package}
\encoding{UTF-8}
\name{restauraR-package}
\alias{restauraR-package}
\alias{restauraR}
\title{restauraR - Restore Ecosystem Services and Biodiversity Through Trait-based Approaches}
\description{
Create restoration solutions to assemble communities from a regional species pool to achieve multiple functional targets based on simulation.
}
\details{
The restauraR package provides an integrative set of tools to create restoration solutions to assemble communities from a regional 
species pool. The framework allows achieving functional targets following multiple approaches: either starting from empty communities or adding species to ongoing 
restoration sites. Detailed methodology is documented in the main function \code{\link{simulateCommunities}} with comprehensive examples available in the package 
vignettes accessible via \code{browseVignettes("restauraR")}.
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Debastiani, V. J., Carlucci, M. B., & Cianciaruso, M. V. (2026). Boosting 
multifunctionality through adaptive trait-based species addition in ongoing restoration projects. 
Ecological Applications, 36, e70197. https://doi.org/10.1002/eap.70197
}
\author{
Vanderlei J. Debastiani <vanderleidebastiani@yahoo.com.br>, André G. Coutinho <andreganemcoutinho@gmail.com>, Marcos B. Carlucci <carlucci@ufpr.br>, Marcus V. Cianciaruso <cianciaruso@gmail.com>
}
