% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAbundance.R, R/sampleAbundanceGroups.R
\encoding{UTF-8}
\name{sampleAbundance}
\alias{sampleAbundance}
\alias{sampleAbundanceGroups}
\title{Internal function to sample species abundance or proportions}
\usage{
sampleAbundance(
  nRich1,
  nRich2,
  sPool,
  nInd,
  cvAbund = 1,
  prob = NULL,
  returnProp = FALSE,
  method = "proportions"
)

sampleAbundanceGroups(
  nRich1,
  nRich2,
  nInd,
  cvAbund = 1,
  prob = NULL,
  returnProp = FALSE,
  method = "proportions",
  group,
  probGroupRich,
  probGroupAbund
)
}
\arguments{
\item{nRich1}{Minimum richness for sample.}

\item{nRich2}{Maximum richness for sample.}

\item{sPool}{Number of species pool.}

\item{nInd}{The number of individuals to draw. Used only in method "individuals".}

\item{cvAbund}{Coefficient of variation (cv) of the relative abundances in the species pool. Used only in method "individuals".}

\item{prob}{Vector of probabilities to draw individuals in each species. Used only in method "individuals".}

\item{returnProp}{Logical argument to specify whether to return proportions of individuals rather than raw abundances.}

\item{method}{Method to obtain the samples, "proportions" or "individuals" (Default method = "proportions").}

\item{group}{A vector with the group to which each species pool belongs.}

\item{probGroupRich}{Vector of probabilities to draw species richness in each group.}

\item{probGroupAbund}{Vector of probabilities to draw individuals or relative abundances in each group.}
}
\value{
A vector with species abundance or proportions.
}
\description{
Internal function to sample species abundance or proportions
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{propMatrix}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
