% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCommunities.R, R/mergeSelection.R,
%   R/print.simRestSelect.R
\encoding{UTF-8}
\name{selectCommunities}
\alias{selectCommunities}
\alias{mergeSelection}
\alias{print.simRestSelect}
\title{Select communities}
\usage{
selectCommunities(x, tests)

mergeSelection(...)

\method{print}{simRestSelect}(x, ...)
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to perform communities selection (or additional selection).}

\item{tests}{A vector with selection criteria to be performed.}

\item{...}{Objects of class "simRestSelect" to be concatenated.}
}
\value{
A list (class "simRestSelect") with the elements:
}
\description{
Select simulated communities based on tests provided
}
