% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeRestauraR.R
\encoding{UTF-8}
\name{themeRestauraR}
\alias{themeRestauraR}
\title{Internal function. A custom theme to use in ggplot2 objects}
\usage{
themeRestauraR(baseSize = 10, legendPosition = "right")
}
\arguments{
\item{baseSize}{Base font size for text elements (default baseSize = 10).}

\item{legendPosition}{Position of the legends (default legendPosition = "left").}
}
\value{
A ggplot theme object.
}
\description{
Internal function. A custom theme to use in ggplot2 objects
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{InternalFunction}
