% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\encoding{UTF-8}
\name{upset}
\alias{upset}
\title{Compose an UpSet plot}
\usage{
upset(
  data,
  intersect,
  base_annotations = "auto",
  name = "group",
  annotations = list(),
  themes = upset_themes(),
  stripes = upset_stripes(),
  labeller = identity,
  height_ratio = 0.5,
  width_ratio = 0.3,
  wrap = FALSE,
  set_sizes = ComplexUpset::upset_set_size(),
  mode = "distinct",
  queries = list(),
  guides = NULL,
  encode_sets = TRUE,
  matrix = ComplexUpset::intersection_matrix(),
  ...
)
}
\arguments{
\item{data}{a dataframe including binary columns representing membership in classes}

\item{intersect}{which columns should be used to compose the intersection}

\item{base_annotations}{a named list with default annotations (i.e. the intersection size barplot)}

\item{name}{the label shown below the intersection matrix}

\item{annotations}{a named list of annotations, each being a list with: `list(aes=mapping, geom=geom or list of geoms)`;
* (optional) `highlight_geom=list of geoms` geoms which can be highlighted with queries,
* (optional) `top_geom=list of geoms` which should show up on top of highlighted queries.}

\item{themes}{a named list of themes for components and annotations, see `upset_default_themes()`/`upset_modify_themes()`}

\item{stripes}{specification of the stripes appearance created with `upset_stripes()`}

\item{labeller}{function modifying the names of the sets (rows in the matrix)}

\item{height_ratio}{ratio of the intersection matrix to intersection size height}

\item{width_ratio}{ratio of the overall set size width to intersection matrix width}

\item{wrap}{whether the plot should be wrapped into a group (makes adding a tile/combining with other plots easier)}

\item{set_sizes}{the overall set sizes plot, e.g. from `upset_set_size()` (`FALSE` to hide)}

\item{mode}{region selection mode for computing the number of elements in intersection fragment. See `get_size_mode()` for accepted values.}

\item{queries}{a list of queries generated with `upset_query()`}

\item{guides}{action for legends aggregation and placement ('keep', 'collect', 'over' the set sizes)}

\item{encode_sets}{whether set names (column in input data) should be encoded as numbers (set to TRUE to overcome R limitations of max 10 kB for variable names for datasets with huge numbers of sets); default TRUE for upset() and FALSE for upset_data().}

\item{matrix}{the intersection matrix plot}

\item{...}{Arguments passed on to `upset_data`}
}
\description{
Modified version of the `upset` function of the `ComplexUpset` package with minimal code correction to work with `ggplot2` 4+
}
\keyword{InternalFunction}
