% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_stripes}
\alias{upset_stripes}
\title{Define appearance of the stripes}
\usage{
upset_stripes(
  mapping = ggplot2::aes(),
  geom = ggplot2::geom_segment(linewidth = 7),
  colors = c("white", "grey95"),
  data = NULL
)
}
\arguments{
\item{mapping}{additional aesthetics}

\item{geom}{a geom to use, should accept `x`, `y`, `xend`, `yend` and `color` aesthetics}

\item{colors}{a vector of colors to repeat as many times as needed for the fill of stripes, or a named vector specifying colors for values of the variable mapped to the color aesthetics in the mapping argument}

\item{data}{the dataset describing the sets with a column named `set` and any other columns as needed for mapping}
}
\description{
Define appearance of the stripes
}
