% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorClass.R
\encoding{UTF-8}
\name{vectorClass}
\alias{vectorClass}
\title{Internal function to return the class of a vector}
\usage{
vectorClass(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The vector class.
}
\description{
Internal function to return the class of a vector
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{InternalFunction}
