% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewMultifunctionality.R
\encoding{UTF-8}
\name{viewMultifunctionality}
\alias{viewMultifunctionality}
\title{function to visualize results}
\usage{
viewMultifunctionality(x, ...)
}
\arguments{
\item{x}{}
}
\description{
Visualize parameters of simulated and selected communities and reference sites
}
