% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjString.R
\encoding{UTF-8}
\name{adjString}
\alias{adjString}
\title{Internal function to format a vector of logical expressions for evaluation within a data frame context}
\usage{
adjString(prefix, logicTest)
}
\arguments{
\item{prefix}{Character string specifying the name of the data frame object.}

\item{logicTest}{The vector containing logical expressions to be evaluated.}
}
\value{
The vector of formatted logical tests.
}
\description{
Internal function to format a vector of logical expressions for evaluation within a data frame context
}
\seealso{
\code{\link{selectCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
