% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCWV.R
\encoding{UTF-8}
\name{calcCWV}
\alias{calcCWV}
\title{Internal function to calculate Community Weighted Variance (CWV)}
\usage{
calcCWV(comm, traits)
}
\arguments{
\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{x}{A community matrix with species composition in the reference sites. NAs not accepted.}
}
\value{
A matrix with Community Weighted Variance.
}
\description{
Implementation is based on the matrix.t function from the SYNCSA package.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
