% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRAO.R
\encoding{UTF-8}
\name{calcRAO}
\alias{calcRAO}
\title{Internal function to calculate Rao's quadratic entropy (RAO)}
\usage{
calcRAO(comm, sppDist = NULL, nRef = NULL, averages = FALSE)
}
\arguments{
\item{comm}{A community matrix with species composition in the reference sites. NAs not accepted.}

\item{sppDist}{Pairwise distance matrix between species (e.g., functional or phylogenetic distances).}

\item{nRef}{Number of reference sites (the first in the species composition matrix) to calculate among-community diversities only partially.}

\item{averages}{Logical argument to specify if return only average for alpha and beta diversities (default averages = FALSE)}
}
\value{
A matrix with among-community diversities excluding within-community diversity.
}
\description{
Efficient implementation to calculate the Rao's quadratic entropy (RAO), based on the discomQE function of the adiv package and on the raoD function of the picante package.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
