% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRAO.R
\encoding{UTF-8}
\name{calcRAO}
\alias{calcRAO}
\title{Internal function to calculate Rao's quadratic entropy (RAO)}
\usage{
calcRAO(comm, dis = NULL, nRef = NULL)
}
\arguments{
\item{comm}{A matrix with species proportions in the reference sites. NAs not accepted.}

\item{dis}{Distances among species.}

\item{nRef}{Number of reference sites (the first in proportions matrix) to calculate among-community diversities only partially}
}
\value{
A matrix with among-community diversities excluding within-community diversity.
}
\description{
Efficient function to calculate the Rao's quadratic entropy (RAO), based on the discomQE function of the adiv package and on raoD function of the picante package.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
