% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateParameters.R
\encoding{UTF-8}
\name{calculateParameters}
\alias{calculateParameters}
\title{calcPar}
\usage{
calculateParameters(
  x,
  trait,
  ava,
  cwm,
  cwv,
  rao,
  cost,
  dens,
  stan,
  reference = NULL,
  supplementary = NULL
)
}
\description{
calcPar
}
