% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerrado.R
\docType{data}
\encoding{UTF-8}
\name{cerrado}
\alias{cerrado}
\title{Cerrado dataset}
\format{
The dataset arranged as a list:
\describe{
  \item{traits}{A data frame object with traits data. The data frame contains four traits: BT (bark thickness), SLA (specific leaf area), Height (maximum height) and Seed (seed mass). The column Available indicates if species are available on the market, the column Cost the cost per individual and the column Density planting density for each species.}
  \item{reference}{A matrix contains relative abundances for 118 species in six reference sites. The species are arranged as columns and sites in rows.}
}
}
\usage{
data(cerrado)
}
\description{
The cerrado dataset contain traits to 483 species used in restoration projects in the savannas in the Cerrado biome.
}
\keyword{datasets}
