% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkReference.R
\encoding{UTF-8}
\name{checkReference}
\alias{checkReference}
\title{checkReference}
\usage{
checkReference(reference, trait, cwm, cwv, rao, stan, supplementary = NULL)
}
\arguments{
\item{reference}{description}

\item{trait}{description}
}
\description{
checkReference
}
