% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkReference.R, R/print.simRestCheck.R
\encoding{UTF-8}
\name{checkReference}
\alias{checkReference}
\alias{print.simRestCheck}
\title{Check species pool and functional parameters in reference sites}
\usage{
checkReference(
  reference,
  trait,
  cwm = NULL,
  cwv = NULL,
  rao = NULL,
  stan = NULL,
  supplementary = NULL,
  props = NULL
)

\method{print}{simRestCheck}(x, ...)
}
\arguments{
\item{reference}{A matrix with species proportions in the reference sites. NAs not accepted.}

\item{trait}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{cwm}{A vector with traits names to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{cwv}{A vector with traits names to calculate Community Weighted Variance (CWV). One CWV is calculated for each trait.}

\item{rao}{A vector with traits names to calculate Rao Quadratic Entropy, or distance matrix (class dist).}

\item{stan}{A vector with parameters names to specify which parameters should be standardized by the maximum.}

\item{supplementary}{A matrix with species proportions in the supplementary sites. NAs not accepted. (default supplementary = NULL).}

\item{props}{Numeric vector of probabilities with values in between 0 and 1 to produces sample quantiles corresponding to the given probabilities (default props = NULL).}

\item{x}{Objects of class "simRestCheck" to print.}

\item{...}{Additional arguments for respective methods.}
}
\value{
A list (class \code{simRestCheck}) with the elements:
\item{call}{The arguments used.}
\item{pool$results}{A data frame with calculated parameters in species pool.}
\item{pool$summary}{A data frame with the descriptive statistics of the species pool.}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{reference$summary}{A data frame with the descriptive statistics of calculated parameters in reference sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
\item{supplementary$summary}{A data frame with the descriptive statistics of calculated parameters in supplementary sites.}
}
\description{
Calculate basic parameters in reference community: richness, Community Weighted Mean, Community Weighted Variance and Rao Quadratic Entropy.
}
\examples{
data("cerrado.mini")
checkReference(reference = cerrado.mini$reference,
               trait = cerrado.mini$traits,
               cwm = "BT",
               rao = c("SLA", "Height", "Seed"),
               props = c(0.75, 0.9))
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{computeParameters}}, \code{\link{selectCommunities}},
\code{\link{extractResults}}, \code{\link{viewResults}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{MainFunction}
