% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkReference.R, R/print.simRestCheck.R
\encoding{UTF-8}
\name{checkReference}
\alias{checkReference}
\alias{print.simRestCheck}
\title{Check species pool and functional parameters in reference sites}
\usage{
checkReference(
  reference,
  traits,
  cwm = NULL,
  cwv = NULL,
  rao = NULL,
  supplementary = NULL,
  props = NULL
)

\method{print}{simRestCheck}(x, ...)
}
\arguments{
\item{reference}{Community matrix with species composition in the reference sites. NAs not accepted.}

\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{cwm}{Character vector of traits names for calculating Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{cwv}{Character vector of traits names for calculating Community Weighted Variance (CWV). One CWV is calculated for each trait.}

\item{rao}{Character vector of traits names for calculating Rao's Quadratic Entropy, a distance matrix (class dist), or a list for calculate multiples Rao.}

\item{supplementary}{Optional community matrix with species composition in the supplementary sites. NAs not accepted. (default supplementary = NULL).}

\item{props}{Numeric vector of probabilities with values in between 0 and 1 to produces sample quantiles corresponding to the given probabilities (default props = NULL).}

\item{x}{An object of class "simRestCheck" to print.}

\item{...}{Additional arguments for respective methods.}
}
\value{
A list (class "simRestCheck") with the elements:
\item{call}{The arguments used.}
\item{pool$results}{Data frame with calculated parameters for species pool.}
\item{pool$summary}{Data frame with the descriptive statistics of the species pool.}
\item{reference$results}{Data frame with calculated parameters for reference sites.}
\item{reference$summary}{Data frame with the descriptive statistics of calculated parameters for reference sites.}
\item{supplementary$results}{Data frame with calculated parameters for supplementary sites.}
\item{supplementary$summary}{Data frame with the descriptive statistics of calculated parameters for supplementary sites.}
}
\description{
Calculate ecological parameters in reference communities: richness, Community Weighted Mean, Community Weighted Variance and Rao's Quadratic Entropy.
}
\examples{
data("cerrado.mini")
checkReference(reference = cerrado.mini$reference,
               traits = cerrado.mini$traits,
               cwm = "BT",
               rao = c("SLA", "Height", "Seed"),
               props = c(0.75, 0.9))
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{computeParameters}}, \code{\link{selectCommunities}},
\code{\link{extractResults}}, \code{\link{viewResults}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{MainFunction}
