% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkResbiotaData.R
\encoding{UTF-8}
\name{checkResbiotaData}
\alias{checkResbiotaData}
\title{Internal function to data validation for resbiota package inputs}
\usage{
checkResbiotaData(
  traits = NULL,
  restComp = NULL,
  restGroup = NULL,
  reference = NULL,
  supplementary = NULL,
  sppDist = NULL,
  cooccur = NULL,
  asList = TRUE
)
}
\arguments{
\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{restComp}{Community matrix with species composition in the restoration sites. NAs not accepted.}

\item{restGroup}{Data frame or matrix with complementary information for restoration sites.}

\item{reference}{Community matrix with species composition in the reference sites.}

\item{supplementary}{Community matrix with species composition in the supplementary sites.}

\item{sppDist}{Pairwise distance matrix between species (e.g., functional or phylogenetic distances).}

\item{cooccur}{A matrix with co-occurrence probabilities between species.}

\item{asList}{A logical argument to specify if return the results in as list format or trigger errors/warnings directly.}
}
\value{
A list with the elements when `asList = TRUE`:
\item{checkStatus}{The global validation status: "success", "warning", or "error".}
\item{checkMessage}{Character vector containing error messages.}
\item{checkWarning}{Character vector containing warning messages.}
}
\description{
Internal function to data validation for resbiota package inputs
}
\seealso{
\code{\link{selectCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{InternalFunction}
