% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeParameters.R, R/computeDissimilarity.R,
%   R/computeMultifunctionality.R
\encoding{UTF-8}
\name{computeParameters}
\alias{computeParameters}
\alias{computeDissimilarity}
\alias{computeMultifunctionality}
\title{Compute functional parameters in communities}
\usage{
computeParameters(
  x,
  trait,
  ava,
  cwm,
  cwv,
  rao,
  cost,
  dens,
  stan,
  reference = NULL,
  supplementary = NULL
)

computeDissimilarity(x, trait)

computeMultifunctionality(x, tests)
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to perform calculate communities parameters.}

\item{trait}{data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{A vector indicating trait name which indicates the availability of species (1 or 0) in trait data.}

\item{cwm}{A vector with traits names to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{cwv}{A vector with traits names to calculate Community Weighted Variance (CWV). One CWV is calculated for each trait.}

\item{rao}{A vector with traits names to calculate Rao Quadratic Entropy, or distance matrix (class dist). Or a list for calculate multiples Rao.}

\item{cost}{A vector with trait name with of species cost per individual}

\item{dens}{A vector with trait name with species planting density}

\item{stan}{A vector with parameters names to specify which parameters should be standardized by the maximum.}

\item{reference}{A matrix with species proportions in the reference sites. NAs not accepted. (default reference = NULL)}

\item{supplementary}{A matrix with species proportions in the supplementary sites. NAs not accepted. (default supplementary = NULL).}

\item{tests}{A vector with multifunctionality criteria to be performed.}
}
\value{
A list (class "simRest" or "simRestSelect") with the elements:
\item{call}{The arguments used.}
\item{simulation$composition}{A matrix with species composition for simulated communities.}
\item{simulation$group}{A data frame with complementary information for restoration sites.}
\item{simulation$results}{A data frame with calculated parameters in each simulated community.}
\item{simulation$multifunctionality}{A binary matrix with multifunctionality tests.}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
}
\description{
Calculate basic parameters in each community: richness, count species unavailable, Community Weighted Mean, Community Weighted Variance, Rao Quadratic Entropy, functional dissimilarity and multifunctionality.
}
\examples{
data("cerrado.mini")
head(cerrado.mini$traits)
# Simulation
scenario <- simulateCommunities(trait = cerrado.mini$traits,
                         ava = "Available",
                         cwm = "BT",
                         rao = c("SLA", "Height", "Seed"),
                         rich = c(10, 15),
                         it = 100)
scenario
# Compute functional parameters
scenario <- computeParameters(x = scenario,
                              trait = cerrado.mini$traits,
                    ava = "Available",
                    cwm = "BT",
                    rao = c("SLA", "Height", "Seed"),
                    cost = "Cost",
                    dens = "Density",
                    reference = cerrado.mini$reference,
                    supplementary = cerrado.mini$supplementary)
scenario
# Compute dissimilarity
scenario <- computeDissimilarity(x = scenario, 
                                 trait = cerrado.mini$traits)
scenario
# Compute multifunctionality
scenario <- computeMultifunctionality(x = scenario, 
                                 tests = c("CWM_BT > 8",
                                           "rao > 2.5"))
scenario
}
\keyword{MainFunction}
