% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeParameters.R,
%   R/computeMultifunctionality.R, R/standardiseParameters.R
\encoding{UTF-8}
\name{computeParameters}
\alias{computeParameters}
\alias{computeMultifunctionality}
\alias{standardiseParameters}
\title{Compute and standardise functional parameters in communities}
\usage{
computeParameters(
  x,
  traits,
  ava = NULL,
  cwm = NULL,
  cwv = NULL,
  rao = NULL,
  cost = NULL,
  dens = NULL,
  traitsFUN = NULL,
  FUN = NULL,
  dissimilarity = NULL,
  reference = NULL,
  supplementary = NULL,
  ...
)

computeMultifunctionality(x, tests)

standardiseParameters(x, parameters = NULL, method = c("max", "standardise"))
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to perform communities parameters.}

\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{Character vector specifying the trait name that indicates the availability of species in traits data (binary: 1 = available, 0 = unavailable).}

\item{cwm}{Character vector specifying traits names to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{cwv}{Character vector specifying traits names to calculate Community Weighted Variance (CWV). One CWV is calculated for each trait.}

\item{rao}{Character vector specifying trait names to calculate Rao's Quadratic Entropy, or distance matrix (class dist). This argument can be a list to calculate multiple Rao indices using different trait sets or species distance matrices.}

\item{cost}{Character vector specifying traits names containing cost per individual for restoration cost estimation.}

\item{dens}{Character vector specifying traits names containing species planting density information for cost calculations. Used only in the method "proportions".}

\item{traitsFUN}{A character vector of trait names to be used in the custom analysis (FUN argument).}

\item{FUN}{A function object to perform the custom analysis.}

\item{dissimilarity}{Character vector specifying traits names to calculate dissimilarity with reference sites, or distance matrix (class dist).}

\item{reference}{Matrix with species composition in the reference sites. NAs not accepted. (default reference = NULL)}

\item{supplementary}{Matrix with species composition in the supplementary sites. NAs not accepted. (default supplementary = NULL).}

\item{...}{Other arguments passed to the custom analysis (FUN argument).}

\item{tests}{Character vector of logical tests for multifunctionality assessment.}

\item{parameters}{Character vector of parameters names to standardised.}

\item{method}{Standardisation method, "max" or "standardise".}
}
\value{
A list (class "simRest" or "simRestSelect") with the elements:
\item{call}{The arguments used.}
\item{simulation$composition}{A matrix with species composition for simulated communities.}
\item{simulation$group}{A data frame with complementary information for restoration sites.}
\item{simulation$baseline}{A matrix with baseline species composition for simulated communities (contains all zeros when restComp is not provided.)}
\item{simulation$results}{A data frame with calculated parameters in each simulated community.}
\item{simulation$multifunctionality}{A data frame with binary multifunctionality tests.}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{reference$multifunctionality}{A data frame with binary multifunctionality tests to reference sites.}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
\item{supplementary$multifunctionality}{A data frame with binary multifunctionality tests to supplementary sites.}
}
\description{
\code{computeParameters} Calculate basic parameters in each simulated community: richness, count species unavailable, Community Weighted Mean, Community Weighted Variance, Rao's quadratic entropy and functional dissimilarity.

\code{standardiseParameters} Standardises the calculated parameters. Two methods are available: "max", which divides the values by the maximum, and "standardise", which scales the calculated parameters to zero mean and unit variance.

\code{computeMultifunctionality} Computes binary multifunctionality matrix and alpha multifunctionality index based on specific test criteria.
}
\examples{
data("cerrado")
head(cerrado$traits)
# Simulation
scenario <- simulateCommunities(traits = cerrado$traits,
                         ava = "Available",
                         maxDiver = c("SLA", "Height", "Seed"),
                         constCWM = "BT",
                         rich = c(10, 15),
                         it = 100)
scenario
# Compute functional parameters
scenario <- computeParameters(x = scenario,
                              traits = cerrado$traits,
                              ava = "Available",
                              cwm = "BT",
                              rao = c("SLA", "Height", "Seed"),
                              cost = "Cost",
                              dens = "Density",
                              dissimilarity = c("SLA", "Height", "Seed"),
                              reference = cerrado$reference)
scenario
# Standardise parameters
scenario <- standardiseParameters(x = scenario,
                                  parameters = "dissimilarity",
                                  method = "max")
scenario
# Compute multifunctionality
scenario <- computeMultifunctionality(x = scenario,
                                 tests = c("CWM_BT > 5.9",
                                           "rao > 0.2"))
scenario
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\seealso{
\code{\link{checkReference}}, \code{\link{simulateCommunities}}, \code{\link{selectCommunities}},
\code{\link{extractResults}}, \code{\link{viewResults}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{MainFunction}
