% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comSimulation.R
\encoding{UTF-8}
\name{comSimulation}
\alias{comSimulation}
\title{function to generate simulated communities}
\usage{
comSimulation(
  trait,
  ava,
  und,
  it,
  rich,
  cwm,
  rao,
  rest,
  max_add,
  min_p,
  phi = 1
)
}
\description{
generates simulated communities and calculate its parameters.
}
