% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\encoding{UTF-8}
\name{extractResults}
\alias{extractResults}
\title{Extract components of simulation results}
\usage{
extractResults(
  x,
  type = "simResults",
  dbFormat = FALSE,
  traits = NULL,
  ava = NULL
)
}
\arguments{
\item{x}{An object of class "simRest" or "simRestSelect" to extract the results.}

\item{type}{An option to extract results, partial match to "simComposition", "simGroup", "simBaseline", "simAdditions", "simResults", "simMultifunctionality", "simMultisiteResults", "simMultisiteCombinations", "simUnavailableSpecies", "refComposition", "refResults", "refMultifunctionality", "supComposition", "supResults", "supMultifunctionality". See value below.}

\item{dbFormat}{Logical argument to specify if return species composition in database format, a data.frame with three columns: "Site", "Species", "Abundance".}

\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{Character vector specifying the trait name that indicates the availability of species in traits data (binary: 1 = available, 0 = unavailable).}
}
\value{
A data.frame or matrix according to the chosen method:
\item{simComposition}{Matrix with species composition for simulated communities.}
\item{simGroup}{Data frame with complementary information for simulated sites.}
\item{simBaseline}{Matrix with baseline species composition for simulated communities.}
\item{simAdditions}{Matrix with species additions for simulated communities.}
\item{simResults}{Data frame with calculated parameters in each simulated community.}
\item{simMultifunctionality}{Data frame with binary multifunctionality tests.}
\item{simMultisiteResults}{Data frame with calculated parameters in each communities combinations.}
\item{simMultisiteCombinations}{Matrix with binary communities combinations.}
\item{simUnavailableSpecies}{Data frame with the species list to achieve the specified thresholds.}
\item{refComposition}{Matrix with species composition for reference sites.}
\item{refResults}{Data frame with calculated parameters in reference sites.}
\item{refMultifunctionality}{Data frame with binary multifunctionality tests for reference sites.}
\item{supComposition}{Matrix with species composition for supplementary sites.}
\item{supResults}{Data frame with calculated parameters in supplementary sites.}
\item{supMultifunctionality}{Data frame with binary multifunctionality tests for supplementary sites.}
}
\description{
Helper function to easily extract specific components from simulation results. Most extraction methods use standard R list indexing. The option "simUnavailableSpecies" (type argument) identifies unavailable species present in selected communities, that is, the species that are necessary to achieve the thresholds but are not available on the market.
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Debastiani, V. J., Carlucci, M. B., & Cianciaruso, M. V. (2026). Boosting 
multifunctionality through adaptive trait-based species addition in ongoing restoration projects. 
Ecological Applications, 36, e70197. https://doi.org/10.1002/eap.70197
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{computeParameters}}, \code{\link{selectCommunities}}, 
\code{\link{viewResults}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{MainFunction}
