% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpecies.R
\encoding{UTF-8}
\name{findSpecies}
\alias{findSpecies}
\title{Internal function to select species to generate simulated communities}
\usage{
findSpecies(traits, maxDiver, constCWM, n, phi)
}
\arguments{
\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{maxDiver}{Character vector specifying traits names to functional diversity optimisation (Rao Quadratic Entropy), or distance matrix (object of class "dist").}

\item{constCWM}{Character vector specifying traits names to constrain Community Weighted Mean (CWM) while maximising functional diversity. Constraints are driven across the range of each trait.}

\item{n}{Number of species to select.}

\item{phi}{Numeric parameter bounded between 0 and 1 that weights the relative importance of either quadratic entropy or entropy.}
}
\value{
A vector with names of selected species, chosen based on their traits and desired trait profile.
}
\description{
Internal function to select species to generate simulated communities
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{generateCommunityMatrices}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
