% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCombinations.R
\encoding{UTF-8}
\name{makeCombinations}
\alias{makeCombinations}
\title{Internal function to make groups combinations}
\usage{
makeCombinations(
  x,
  groups = NULL,
  minSubset = 1,
  maxSubset = 1,
  maxComb = 1000
)
}
\arguments{
\item{x}{A vector with elements from which to generate the combinations.}

\item{groups}{A vector with the groups which the element belongs.}

\item{minSubset}{Minimum number of elements to include in each combination.}

\item{maxSubset}{Maximum number of elements to include in each combination.}

\item{maxComb}{Maximum number of combinations to generate.}
}
\value{
A binary matrix with all elements combinations.
}
\description{
Create element combinations with group-based constraints.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{InternalFunction}
