% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multCalculation.R
\encoding{UTF-8}
\name{multCalculation}
\alias{multCalculation}
\title{function to calculate multifunctionality}
\usage{
multCalculation(x, tests, where = "global")
}
\arguments{
\item{x}{x}

\item{tests}{tests}

\item{where}{where}
}
\description{
function to calculate multifunctionality
}
