% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propMatrix.R
\encoding{UTF-8}
\name{propMatrix}
\alias{propMatrix}
\title{Internal function to generate community matrices}
\usage{
propMatrix(
  traits,
  ava,
  und,
  it,
  rich,
  maxDiver,
  constCWM,
  phi,
  nInd,
  cvAbund,
  prob,
  method,
  cooccur,
  minAbund,
  group,
  probGroupRich,
  probGroupAbund
)
}
\arguments{
\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{Character vector specifying trait name which indicates the availability of species in traits data (binary: 1 = available, 0 = unavailable)..}

\item{und}{Character vector specifying trait name which indicates undesired species in traits data (binary: 1 = undesired, 0 = desired).}

\item{it}{Number of iterations (communities).}

\item{rich}{Numeric vector specifying the richness range in each community.}

\item{maxDiver}{A vector of traits names to functional diversity optimisation (Rao Quadratic Entropy), or distance matrix (object of class "dist").}

\item{constCWM}{A vector of trait names to constrain Community Weighted Mean (CWM) while maximising functional diversity. Constraints are driven across the range of each trait.}

\item{phi}{Numeric parameter bounded between 0 and 1 that weights the relative importance of either quadratic entropy or entropy.}

\item{nInd}{Numeric vector with the number of individuals to draw. Used only in method "individuals".}

\item{cvAbund}{Coefficient of variation (cv) of the relative abundances in the species pool. Used only with method "individuals".}

\item{prob}{Character vector specifying trait name which indicates the probabilities to draw individuals in each species. Used only in method "individuals".}

\item{method}{Method to obtain the samples, "proportions" or "individuals" (Default method = "proportions").}

\item{cooccur}{Matrix with co-occurrence probabilities between species to constrain ecological assembly patterns.}

\item{minAbund}{Minimal abundance or proportion threshold to maintain for each species in simulated communities.}

\item{group}{Character vector specifying the trait name which indicates the group to which the species belongs.}

\item{probGroupRich}{Numeric vector of probabilities to draw species richness in each group.}

\item{probGroupAbund}{Numeric vector of probabilities to draw individuals or relative abundances in each group.}
}
\value{
A community matrix with sites as rows and species as columns, containing either relative species proportions or raw species abundances counts.
}
\description{
Internal function to generate community matrices
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{findSpecies}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
