% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propMatrix.R
\encoding{UTF-8}
\name{propMatrix}
\alias{propMatrix}
\title{Internal function to generate communities matrix}
\usage{
propMatrix(
  trait,
  ava,
  und,
  it,
  rich,
  cwm,
  rao,
  phi,
  nInd,
  cvAbund,
  prob,
  method,
  group,
  probGroupRich,
  probGroupAbund
)
}
\arguments{
\item{trait}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{ava}{A vector indicating trait name which indicates the availability of species (1 or 0) in trait data.}

\item{und}{A vector indicating trait name which indicates undesired species (1 or 0) in trait data.}

\item{it}{Number of iterations (communities).}

\item{rich}{The range of richness values in each community.}

\item{cwm}{A vector with trait names to constrain Community Weighted Mean (CWM) while maximising functional diversity. Constraints are driven over the range of each trait.}

\item{rao}{A vector with traits names to be considered in maximize functional diversity (Rao Quadratic Entropy), or distance matrix (class "dist").}

\item{phi}{A parameter bounded between 0 and 1 that weights the importance of either quadratic entropy or entropy.}

\item{nInd}{The number of individuals to draw. Used only in method "individuals".}

\item{cvAbund}{Coefficient of variation (cv) of the relative abundances in the species pool. Used only in method "individuals".}

\item{prob}{A vector indicating trait name which indicates the probabilities to draw individuals in each species. Used only in method "individuals".}

\item{method}{Method to obtain the samples, "proportions" or "individuals" (Default method = "proportions").}

\item{group}{A vector with traits name which indicates the group to which species belongs.}

\item{probGroupRich}{Vector of probabilities to draw species richness in each group.}

\item{probGroupAbund}{Vector of probabilities to draw individuals or relative abundances in each group.}
}
\value{
A community matrix with relative or raw species abundances.
}
\description{
Internal function to generate communities matrix
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{findSpecies}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
