% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randDivision.R
\encoding{UTF-8}
\name{randDivision}
\alias{randDivision}
\title{Internal function to randomly divide an integer into groups}
\usage{
randDivision(n, k)
}
\arguments{
\item{n}{An integer to be divided.}

\item{k}{Number of groups.}
}
\value{
A numeric vector with the random division.
}
\description{
Internal function to randomly divide an integer into groups
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{generateCommunityMatrices}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{InternalFunction}
