% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readjustAbundance.R
\encoding{UTF-8}
\name{readjustAbundance}
\alias{readjustAbundance}
\title{Internal function to randomly ajust species abundance or proportions}
\usage{
readjustAbundance(abund0, minAbund, method)
}
\arguments{
\item{abund0}{A vector with species abundance or proportions to adjust.}

\item{minAbund}{Minimal abundance or proportion to keep in simulated communities.}

\item{method}{The method used to obtain the samples, "proportions" or "individuals".}
}
\value{
A vector with species abundance or proportions adjusted.
}
\description{
Internal function to randomly ajust species abundance or proportions
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{propMatrix}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
