% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangementMatrix.R, R/makeMatrixTemplate.R
\encoding{UTF-8}
\name{rearrangementMatrix}
\alias{rearrangementMatrix}
\alias{makeMatrixTemplate}
\title{Internal function to rearrange matrix structures}
\usage{
rearrangementMatrix(template, x, fillNA = FALSE)

makeMatrixTemplate(...)
}
\arguments{
\item{template}{A template matrix returned by \code{makeMatrixTemplate} function to guide rearrangement.}

\item{x}{A data.frame or matrix to be reorganised.}

\item{fillNA}{Logical argument (TRUE or FALSE) to specify if missing cells (NA) should be replaced with zeros (default fillNA = FALSE).}

\item{...}{Objects of class data.frame (or matrix) to be concatenated.}
}
\value{
The \code{makeMatrixTemplate} function returns an empty data.table object with a reference structure, while \code{rearrangementMatrix} aligns existing data to that structure and returns the rearranged matrix.
}
\description{
\code{makeMatrixTemplate} Creates an empty template matrix. This function concatenates two or more data.frames (or matrices) 
and returns an empty data.table object (template) with 0 rows and column names equal to the union of all unique column names from the input objects. 

\code{rearrangementMatrix} rearranges one matrix following a template layout. The names of the columns are required to rearrange the matrices.
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{InternalFunction}
