% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizeMatrix.R, R/makeMatrixTemplate.R
\encoding{UTF-8}
\name{reorganizeMatrix}
\alias{reorganizeMatrix}
\alias{makeMatrixTemplate}
\title{Internal function to reorganize matrix}
\usage{
reorganizeMatrix(template, x, fillNA = FALSE)

makeMatrixTemplate(...)
}
\arguments{
\item{template}{A template matrix returned by the \code{makeMatrixTemplate} function.}

\item{x}{A data.frame or matrix to be reorganized.}

\item{fillNA}{Logical argument (TRUE or FALSE) to specify if missing cells with NAs are replaced with zero (0).}

\item{...}{Objects of class data.frame (or matrix) to be concatenated.}
}
\value{
The \code{makeMatrixTemplate} function returns a empty data.table object, \code{reorganizeMatrix} returns the reorganized matrix.
}
\description{
\code{makeMatrixTemplate} creates a template matrix. This function concatenates two or more data.frame (or matrix) 
and return an empty data.table object (template) with 0 rows and columns equal to all unique columns of the input objects. 

\code{reorganizeMatrix} rearranges one matrix following a template layout. The names of the columns are required to reorganise the matrices.
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
