% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDivision.R
\encoding{UTF-8}
\name{roundDivision}
\alias{roundDivision}
\title{Internal function to divide an integer into proportional groups}
\usage{
roundDivision(n, prop)
}
\arguments{
\item{n}{An integer to be divided.}

\item{prop}{A numeric vector with proportions to divide for each group.}
}
\value{
An integer vector with distribution across groups. Any remaining part is randomly distributed among groups.
}
\description{
Internal function to divide an integer into proportional groups
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{propMatrix}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
