% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDivision.R
\encoding{UTF-8}
\name{roundDivision}
\alias{roundDivision}
\title{Internal function to return only integers in proportional division.}
\usage{
roundDivision(n, prop)
}
\arguments{
\item{n}{A integer to divide.}

\item{prop}{A numeric vector with proportions.}
}
\value{
An integer vector with division. The remainder parts are randomly distributed in the groups.
}
\description{
Internal function to return only integers in proportional division.
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{propMatrix}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
