% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAbundance.R, R/sampleAbundanceGroups.R
\encoding{UTF-8}
\name{sampleAbundance}
\alias{sampleAbundance}
\alias{sampleAbundanceGroups}
\title{Internal function to sample species abundance or proportions}
\usage{
sampleAbundance(
  nRich1,
  nRich2,
  sPool,
  nInd1,
  nInd2,
  cvAbund = 1,
  prob = NULL,
  returnProp = FALSE,
  method = "proportions",
  cooccur = NULL,
  minAbund = NULL
)

sampleAbundanceGroups(
  nRich1,
  nRich2,
  nInd1,
  nInd2,
  cvAbund = 1,
  prob = NULL,
  returnProp = FALSE,
  method = "proportions",
  cooccur = NULL,
  minAbund = NULL,
  group,
  probGroupRich,
  probGroupAbund
)
}
\arguments{
\item{nRich1}{Minimum species richness to sample.}

\item{nRich2}{Maximum species richness to sample.}

\item{sPool}{Total number of species in species pool.}

\item{nInd1}{Minimum number of individuals to draw. Used only in the method "individuals".}

\item{nInd2}{Maximum number of individuals to draw. Used only in the method "individuals".}

\item{cvAbund}{Coefficient of variation (cv) of the relative abundances in the species pool. Used only in the method "individuals".}

\item{prob}{Vector of probabilities to draw individuals in each species. Used only in the method "individuals".}

\item{returnProp}{Logical argument to specify whether to return proportions of individuals rather than raw abundances.}

\item{method}{Method to obtain the samples, "proportions" or "individuals" (Default method = "proportions").}

\item{cooccur}{A matrix with co-occurrence probabilities between species.}

\item{minAbund}{Minimal abundance or proportion to keep in simulated communities.}

\item{group}{Character vector to specify group membership for each species.}

\item{probGroupRich}{Vector of probabilities to draw species richness in each group.}

\item{probGroupAbund}{Vector of probabilities to draw individuals or relative abundances in each group.}
}
\value{
A vector with species abundance or proportions.
}
\description{
Internal function to sample species abundance or proportions
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{generateCommunityMatrices}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{InternalFunction}
