% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCommunities.R, R/mergeSelection.R,
%   R/print.simRestSelect.R
\encoding{UTF-8}
\name{selectCommunities}
\alias{selectCommunities}
\alias{mergeSelection}
\alias{print.simRestSelect}
\title{Select communities}
\usage{
selectCommunities(x, tests)

mergeSelection(...)

\method{print}{simRestSelect}(x, ...)
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to perform communities selection (or additional selection).}

\item{tests}{A vector with selection criteria to be performed.}

\item{...}{Objects of class "simRestSelect" to be concatenated.}
}
\value{
A list (class "simRestSelect") with the elements:
\item{call}{The arguments used.}
\item{selection$composition}{A matrix with species composition for selected communities.}
\item{selection$group}{A data frame with complementary information for selected sites.}
\item{selection$results}{A data frame with calculated parameters in each selected community.}
\item{selection$N}{FINALIZAR}
\item{selection$thresholds}{FINALIZAR}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
}
\description{
Select simulated communities based on tests provided
}
\examples{
data("cerrado.mini")
head(cerrado.mini$traits)
# Simulation
scenario <- simulateCommunities(trait = cerrado.mini$traits,
                         ava = "Available",
                         cwm = "BT",
                         rao = c("SLA", "Height", "Seed"),
                         rich = c(10, 15),
                         it = 100)
scenario
# Compute functional parameters
scenario <- computeParameters(x = scenario,
                              trait = cerrado.mini$traits,
                    ava = "Available",
                    cwm = "BT",
                    rao = c("SLA", "Height", "Seed"),
                    cost = "Cost",
                    dens = "Density",
                    reference = cerrado.mini$reference,
                    supplementary = cerrado.mini$supplementary)
scenario
# Select communities
scenarioSelected <- selectCommunities(x = scenario,
                                      tests = c("CWM_BT > 8",
                                                "rao > 2.5"))
scenarioSelected
}
\keyword{MainFunction}
