% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCommunities.R
\encoding{UTF-8}
\name{selectCommunities}
\alias{selectCommunities}
\title{function to select communities (comSelection)}
\usage{
selectCommunities(x, tests)
}
\arguments{
\item{tests}{list with tests to be performed}

\item{param}{data.frame with parameters of simulated communities}

\item{comp}{data.frame with compositions of simulated communities}
}
\description{
select simulated communities based on tests provided and add information to x
}
