% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCommunities.R, R/mergeSelection.R,
%   R/print.simRestSelect.R
\encoding{UTF-8}
\name{selectCommunities}
\alias{selectCommunities}
\alias{mergeSelection}
\alias{print.simRestSelect}
\title{Select communities}
\usage{
selectCommunities(
  x,
  testsDet = NULL,
  testsHie = NULL,
  group = NULL,
  singleselection = TRUE
)

mergeSelection(...)

\method{print}{simRestSelect}(x, ...)
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to perform communities selection (or additional selection). Or an object of class "simRestSelect" to print.}

\item{testsDet}{A vector with the deterministic selection criteria to be executed.}

\item{testsHie}{A vector with the hierarchical selection criteria to be executed.}

\item{group}{A vector with a parameter name to specify the simulation groups. This is only used for the hierarchical selection.}

\item{singleselection}{A logical argument to specify if only one simulation is selected by group (default singleselection = TRUE). This is only used for the hierarchical selection.}

\item{...}{Objects of class "simRestSelect" to be concatenated. Additional arguments for respective methods.}
}
\value{
A list (class "simRestSelect") with the elements:
\item{call}{The arguments used.}
\item{selection$composition}{A matrix with species composition for selected communities.}
\item{selection$group}{A data frame with complementary information for selected sites.}
\item{selection$results}{A data frame with calculated parameters in each selected community.}
\item{selection$multifunctionality}{A data frame with binary multifunctionality tests.}
\item{selection$thresholds}{A vector with the count of selected communities at each threshold. When simulations are merged, it is not shown.}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{reference$multifunctionality}{A data frame with binary multifunctionality tests to reference sites.}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
\item{supplementary$multifunctionality}{A data frame with binary multifunctionality tests to supplementary sites.}
}
\description{
Select simulated communities based on tests provided
}
\examples{
data("cerrado.mini")
head(cerrado.mini$traits)
# Simulation
scenario <- simulateCommunities(trait = cerrado.mini$traits,
                         ava = "Available",
                         cwm = "BT",
                         rao = c("SLA", "Height", "Seed"),
                         rich = c(10, 15),
                         it = 100)
scenario
# Compute functional parameters
scenario <- computeParameters(x = scenario,
                              trait = cerrado.mini$traits,
                    ava = "Available",
                    cwm = "BT",
                    rao = c("SLA", "Height", "Seed"),
                    cost = "Cost",
                    dens = "Density",
                    reference = cerrado.mini$reference,
                    supplementary = cerrado.mini$supplementary)
scenario
# Select communities - Deterministic selection
scenarioSelected <- selectCommunities(x = scenario,
                                      testsDet = c("CWM_BT > 6",
                                                "rao > 2.5"))
scenarioSelected
# Select communities - Hierarchical selection
scenarioSelected <- selectCommunities(x = scenario,
                                      testsHie = c("CWM_BT > 6",
                                                "rao > 2.5",
                                                "cost == 'MIN'"))
scenarioSelected
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\seealso{
\code{\link{checkReference}}, \code{\link{simulateCommunities}}, \code{\link{computeParameters}},
\code{\link{extractResults}}, \code{\link{viewResults}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{MainFunction}
