% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCommunities.R, R/mergeSelection.R,
%   R/print.simRestSelect.R, R/print.summarySimRestSelect.R,
%   R/summary.simRestSelect.R
\encoding{UTF-8}
\name{selectCommunities}
\alias{selectCommunities}
\alias{mergeSelection}
\alias{print.simRestSelect}
\alias{print.summarySimRestSelect}
\alias{summary.simRestSelect}
\title{Select communities}
\usage{
selectCommunities(
  x,
  testsFilter = NULL,
  testsPriority = NULL,
  siteGroup = NULL,
  singleSelection = FALSE,
  testsMultisite = NULL
)

mergeSelection(...)

\method{print}{simRestSelect}(x, ...)

\method{print}{summarySimRestSelect}(x, ...)

\method{summary}{simRestSelect}(object, ...)
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to perform communities selection (or additional selection). Or an object of class "simRestSelect" to print.}

\item{testsFilter}{Character vector of logical tests for filter selection. Communities must satisfy all specified conditions to be selected.}

\item{testsPriority}{Character vector of logical tests for hierarchical priority selection. Tests are evaluated in order until at least one community is selected.}

\item{siteGroup}{Character vector specifying a parameter name to define site groups for priority selection. This is only used for the priority selection.}

\item{singleSelection}{Logical argument to specify if only one simulation is selected by site group (default singleSelection = FALSE). This is only used for the priority selection.}

\item{testsMultisite}{Character vector of logical tests for multi-site selection.}

\item{...}{Objects of class "simRestSelect" to be concatenated. Additional arguments for respective methods.}

\item{object}{Object of class "simRestSelect" to summarise.}
}
\value{
A list (class "simRestSelect") with the elements:
\item{call}{The arguments used.}
\item{selection$composition}{A matrix with species composition for selected communities.}
\item{selection$group}{A data frame with complementary information for selected sites.}
\item{selection$baseline}{A matrix with with baseline species composition for simulated communities (contains all zeros when restComp is not provided.)}
\item{selection$results}{A data frame with calculated parameters in each selected community.}
\item{selection$multifunctionality}{A data frame with binary multifunctionality tests.}
\item{selection$thresholds}{A vector with the count of selected communities at each threshold. When simulations are merged, it is not shown.}
\item{selection$multisite$combinations}{A binary matrix with communities combinations.}
\item{selection$multisite$results}{A data frame with calculated parameters in each communities combinations.}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{reference$multifunctionality}{A data frame with binary multifunctionality tests to reference sites.}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
\item{supplementary$multifunctionality}{A data frame with binary multifunctionality tests to supplementary sites.}
}
\description{
Select simulated communities using filter-based, priority-based, or multi-site selection methods to identify optimal restoration solutions.
}
\examples{
data("cerrado.mini")
head(cerrado.mini$traits)
# Simulation
scenario <- simulateCommunities(traits = cerrado.mini$traits,
                                ava = "Available",
                                maxDiver = c("SLA", "Height", "Seed"),
                                constCWM = "BT",
                                rich = c(10, 15),
                                it = 100)
scenario
# Compute functional parameters
scenario <- computeParameters(x = scenario,
                              traits = cerrado.mini$traits,
                              ava = "Available",
                              cwm = "BT",
                              rao = c("SLA", "Height", "Seed"),
                              cost = "Cost",
                              dens = "Density",
                              reference = cerrado.mini$reference)
scenario
# Select communities - Filter selection
scenarioSelectedFilter <- selectCommunities(x = scenario,
                                            testsFilter = c("CWM_BT > 5.9",
                                                            "rao > 0.2"))
scenarioSelectedFilter
# Select communities - Priority selection
scenarioSelectedPriority <- selectCommunities(x = scenario,
                                              testsPriority = c("CWM_BT > 5.9",
                                                                "rao > 0.2",
                                                                "Cost == 'MIN'"))
scenarioSelectedPriority
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\seealso{
\code{\link{checkReference}}, \code{\link{simulateCommunities}}, \code{\link{computeParameters}},
\code{\link{extractResults}}, \code{\link{viewResults}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
\keyword{MainFunction}
