% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCommunities.R, R/mergeSimulations.R,
%   R/print.simRest.R
\encoding{UTF-8}
\name{simulateCommunities}
\alias{simulateCommunities}
\alias{mergeSimulations}
\alias{print.simRest}
\title{Generate simulated communities to restoration}
\usage{
simulateCommunities(
  trait,
  restComp,
  restGroup,
  ava,
  und,
  it,
  rich,
  cwm,
  rao,
  max_add,
  min_p,
  phi = 1,
  prefix = NULL
)

mergeSimulations(...)

\method{print}{simRest}(x, ...)
}
\arguments{
\item{trait}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{restComp}{A matrix with species proportions in the restoration sites. NAs not accepted.}

\item{restGroup}{Data frame or matrix with complementary information for restoration sites.}

\item{ava}{A vector indicating trait name which indicates the availability of species (1 or 0) in trait data.}

\item{und}{A vector indicating trait name which indicates undesired species (1 or 0) in trait data.}

\item{it}{Number of iterations (communities).}

\item{rich}{The range of richness values in each community.}

\item{cwm}{A vector with traits names to calculate Community Weighted Mean (CWM). One CWM is calculated for each trait.}

\item{rao}{A vector with traits names to calculate Rao Quadratic Entropy, or distance matrix (class dist).}

\item{phi}{A parameter bounded between 0 and 1 that weights the importance of either quadratic entropy or entropy (default phi = 1).}

\item{prefix}{A prefix to use in current simulation.}

\item{...}{Objects of class "simRest" to be concatenated.}
}
\value{
A list (class "simRest") with the elements:
\item{call}{The arguments used.}
\item{simulation$composition}{A matrix with species composition for simulated communities.}
\item{simulation$group}{A data frame with complementary information for restoration sites.}
\item{simulation$results}{A data frame with calculated parameters in each simulated community.}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{reference$results}{A data frame with calculated parameters in reference sites.}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
\item{supplementary$results}{A data frame with calculated parameters in supplementary sites.}
}
\description{
Generates simulated communities.
}
\examples{
data("cerrado.mini")
head(cerrado.mini$traits)
# Restoration new sites
scenarioA <- simulateCommunities(trait = cerrado.mini$traits,
                         ava = "Available",
                         cwm = "BT",
                         rao = c("SLA", "Height", "Seed"),
                         rich = c(10, 15),
                         it = 100)
scenarioA
# Restoration existing sites
scenarioB <- simulateCommunities(trait = cerrado.mini$traits,
                    restComp = cerrado.mini$restoration,
                    ava = "Available",
                    cwm = "BT",
                    rao = c("SLA", "Height", "Seed"),
                    rich = c(10, 15),
                    it = 100,
                    max_add = 10)
scenarioB
# Merge all scenarios
allScenarios <- mergeSimulations(scenarioA, scenarioB)
allScenarios
}
\keyword{MainFunction}
