% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCommunities.R
\encoding{UTF-8}
\name{simulateCommunities}
\alias{simulateCommunities}
\title{function to generate simulated communities (comSimulation)}
\usage{
simulateCommunities(
  trait,
  restComp,
  restGroup,
  ava,
  und,
  it,
  rich,
  cwm,
  rao,
  max_add,
  min_p,
  phi = 1,
  prefix = NULL
)
}
\description{
generates simulated communities and calculate its parameters.
}
