% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCommunities.R, R/mergeSimulations.R,
%   R/print.simRest.R, R/print.summarySimRest.R, R/summary.simRest.R
\encoding{UTF-8}
\name{simulateCommunities}
\alias{simulateCommunities}
\alias{mergeSimulations}
\alias{print.simRest}
\alias{print.summarySimRest}
\alias{summary.simRest}
\title{Simulate ecological communities for restoration planning}
\usage{
simulateCommunities(
  traits,
  restComp = NULL,
  restGroup = NULL,
  ava = NULL,
  und = NULL,
  it = 1000,
  rich,
  maxDiver = NULL,
  constCWM = NULL,
  prob = NULL,
  phi = 1,
  nInd = NULL,
  cvAbund = 1,
  prefix = NULL,
  method = "proportions",
  cooccur = NULL,
  minAbund = NULL,
  group = NULL,
  probGroupRich = NULL,
  probGroupAbund = NULL
)

mergeSimulations(...)

\method{print}{simRest}(x, ...)

\method{print}{summarySimRest}(x, ...)

\method{summary}{simRest}(object, ...)
}
\arguments{
\item{traits}{Data frame or matrix with species traits. Traits as columns and species as rows.}

\item{restComp}{Matrix with species composition for restoration sites. NAs not accepted.}

\item{restGroup}{Data frame or matrix with complementary site-specific information for restoration sites.}

\item{ava}{Character vector specifying the trait name that indicates the availability of species in traits data (binary: 1 = available, 0 = unavailable).}

\item{und}{Character vector specifying the trait name that indicates undesired species in traits data (binary: 1 = undesired, 0 = desired).}

\item{it}{Number of iterations (communities) to generate.}

\item{rich}{Numeric vector specifying the richness range in each community, or a character vector specifying the restGroup columns that indicate the richness range within site-specific restoration targets.}

\item{maxDiver}{Character vector specifying trait names to functional diversity optimisation (Rao Quadratic Entropy), or distance matrix (object of class "dist").}

\item{constCWM}{Character vector specifying trait names to constrain Community Weighted Mean (CWM) while maximising functional diversity. Constraints are driven across the range of each trait.}

\item{prob}{Character vector specifying the trait name that indicates the probabilities to draw individuals in each species. Used only in the method "individuals".}

\item{phi}{Numeric parameter bounded between 0 and 1 that weights the relative importance of either quadratic entropy or entropy.}

\item{nInd}{Numeric vector with the number of individuals to draw, or a character vector specifying the restGroup column that indicates the number of individuals to draw within site-specific restoration targets. Used only in the method "individuals".}

\item{cvAbund}{Coefficient of variation (cv) of the relative abundances in the species pool. Used only in the method "individuals".}

\item{prefix}{Character prefix for naming simulation outputs.}

\item{method}{Method to obtain the samples, "proportions" or "individuals" (Default method = "proportions").}

\item{cooccur}{Matrix with co-occurrence probabilities between species to constrain ecological assembly patterns.}

\item{minAbund}{Minimal abundance or proportion threshold to maintain for each species in simulated communities.}

\item{group}{Character vector specifying the trait name that indicates the group to which the species belongs.}

\item{probGroupRich}{Numeric vector of probabilities to draw species richness in each group.}

\item{probGroupAbund}{Numeric vector of probabilities to draw individuals or relative abundances in each group.}

\item{...}{Objects of class "simRest" to be concatenated. Additional arguments for respective methods.}

\item{x}{Object of class "simRest" to print.}

\item{object}{Object of class "simRest" to summarise.}
}
\value{
A list (class "simRest") with the elements:
\item{call}{The arguments used.}
\item{simulation$composition}{A matrix with species composition for simulated communities.}
\item{simulation$group}{A data frame with complementary information for restoration sites.}
\item{simulation$baseline}{A matrix with with baseline species composition for simulated communities (contains all zeros when restComp is not provided.)}
\item{reference$composition}{A matrix with species composition for reference sites}
\item{supplementary$composition}{A matrix with species composition for supplementary sites.}
}
\description{
Generates simulated communities for ecological restoration using trait-based approaches with multiple ecological constraints.
}
\details{
\strong{Overview}

The framework generates communities with distinct species compositions drawn from a species pool. The algorithm takes as input a data set with trait information for the species in the regional pool, that is, a list of species from which trait data are available that could be used to restore a particular ecosystem. Using species from the regional pool, the algorithm produces a set of N-simulated communities with species richness that ranges within a user-defined interval. 

\strong{Workflow}

The basic workflow includes:
\itemize{
  \item Set simulation input parameters and do the simulation (\code{simulateCommunities});
  \item Compute and standardise parameters in each simulated community (\code{computeParameters}, \code{standardiseParameters});
  \item Calculate multiple restoration targets (\code{computeMultifunctionality});
  \item Selections of simulated communities (\code{selectCommunities});
  \item Views, extract and save the results (\code{viewResults}, \code{viewMultifunctionality}, \code{extractResults}).
}

Note that (\code{simulateCommunities}) returns communities without calculated functional parameters. Use (\code{computeParameters}) to calculate CWM, Rao's, etc., before selection.

\strong{Traits and additional species information}

Each species in the pool can be described by functional traits that represent the ecological functions of communities, ecosystem services, or desirable characteristics for novel or restored communities. Those species can be classified into species that are available on the market and those that are not. Furthermore, the cost per individual, observed frequencies occurring in reference sites, and any additional information can be used to select species composition.

\strong{Simulations}

The species assemblages are generated in the \code{simulateCommunities} function. The species richness within simulated communities varies within a range defined by the user. Thus, it could be specified using the observed range in reference communities or desired restoration targets. 

The sampling of species can be performed completely at random or select species based on functional traits or a desired trait profile. If the `maxDiver` argument is specified, species selections optimise the functional diversity of a specific set of traits. Simultaneously, the `constCWM` argument can constrain the community-weighted mean trait values across the full range of each trait enabling local maximization of functional diversity within the specified community-weighted mean constraints.

To obtain a high number of restoration solutions, all runs of the \code{simulateCommunities} function return a set of simulated communities with random species distribution, and composition based on rules set with `maxDiver` and `constCWM` arguments. Furthermore, if the species available is informed in the `ava` argument, the sample of species compositions is performed also with only available species on the market. Thus, the function returns a wide range of variability in composition to posterior selection, including compositions with available and not available species in the market.

The framework returns a community matrix indicating the composition of individuals that need to be added for each species. If no established communities are informed, the simulated communities are set as empty communities (sites to restore start with no species, and all species must be planted for restoration). Alternatively, if established communities are set, the new species and individuals are introduced into the established communities (sites to restore can start with pre-existing species). Thus, it is possible to increase the number of ecosystem functions in ongoing restoration sites and follow an approach of adaptive management.

\strong{Communities parameters}

After the initial step of species composition simulation, several functional parameters can be calculated for each community. In the restauraR package, this step is mandatory, given that those parameters are used to select a subset of simulated communities that fit the restoration target. Parameters such as species richness, Functional Diversity (FD), Community Weighted Means (CWM), the average functional dissimilarity between each simulated community, and restoration cost can be calculated. In general, those function parameters are used as a proxy for ecosystem functions relevant to the restoration under evaluation. The functions \code{computeParameters} and \code{computeMultifunctionality} can be used in this step. 

The function \code{computeParameters} is mandatory and computes basic parameters: richness, Community Weighted Mean, Community Weighted Variance, Rao’s Quadratic Entropy and functional dissimilarity. The user must define which traits will be used to calculate each metric. Additionally, if species cost per individual and planting density information are provided, restoration costs are estimated. When the availability of species is specified, the function counts species unavailable in each community. Furthermore, it is optional to compute functional dissimilarity between selected communities and reference sites. Functional dissimilarity is a pair-wise metric between each selected community and each reference site. If you have more than one reference site, functional dissimilarity will be the average dissimilarity between each selected community and the reference sites. Functional dissimilarity is calculated with the function discomQE from the package adiv using the default formula “QE”. Therefore, it corresponds to the Rao’s dissimilarity between communities. 

The function \code{standardiseParameters} allows the standardisation of the calculated parameters. Two methods are available: "max", which divides the values by the maximum, and "standardise", which scales the calculated parameters to zero mean and unit variance. For example, the "max" method is useful to scale dissimilarity results by scaling the values in ranges from 0 (no dissimilarity) to 1 (maximum dissimilarity).

The function \code{computeMultifunctionality} is optional and computes the matrix of multifunctionality and alpha multifunctionality. See the multifunctionality topic below.

\strong{Select communities}

The selections of simulated communities are performed using the function \code{selectCommunities}. Selections are based on simple logical tests based on thresholds. The user must define parameter thresholds that indicate the level of functional composition above (or below) which the function is considered to be restored in each site. The thresholds can be set for one or more calculated parameters in the previous step. This is an essential step that must be based on previous studies that indicate the relationship between species traits and ecosystem functions related to the desired restoration targets. Knowledge of the regional species pool, species availability and reference sites are important to find solutions that adequately optimise restoration goals.

The logical tests are specified using a vector of character class. Each test uses the parameter name, a logical operator and the value threshold. All basic logical tests can be used, setting the default operators: <, <=, >, >=, == and !=.  Each test must be in double or single quote signs. The selection of character type can be performed using different quotation marks than those used to define the entire test (e.g., single quote inside the double quote).

Two selection options can be used, individually or interchangeably: filter selection and priority selection. In the filter selection, all simulations that satisfy all the provided test conditions are returned. In this case, all parameters evaluated must be true for a community to be selected. Multiple selection steps can be performed sequentially, using more restrictive criteria (although the practical results will be the same when applied in a single step with the most restrictive criterion). In the priority selection, the tests are evaluated hierarchically following an order until a single simulation is selected. When all simulations fail in the first test, the function tries the next test. In the end, the function samples only one simulation from among those that passed all the tests. In this case, an additional argument defines if the selection is performed inside a specific group. This allows, for example, the selection of one simulation for each site that will be restored. Furthermore, the priority selection supports two special keywords: MIN to select the minimal value for a specified parameter, and MAX to select the maximum value.

\strong{The reference sites}

Reference sites can be included in the \code{computeParameters} step. Thus, all calculated parameters are also calculated for reference sites and can be used in posterior analysis or as an auxiliary to the selection procedure by identifying the natural range and values of these parameters in the reference ecosystem. 

The restauraR package also allows the inclusion of supplementary sites to calculate the same set of functional parameters. These supplementary sites are only used to compare proposals, for example, to evaluate the parameters in naturally regenerated species sites.

\strong{The outputs}

The main outputs are a community matrix with species relative or raw abundances and a data frame with the functional parameters calculated for each community. These outputs allow the user to investigate the relationship among these parameters, select communities that meet the restoration goals, and identify functionally important species that are not available on the market. 

\strong{Multifunctionality}

Restoration projects can target restoring multiple ecosystem services, called multifunctionality. Thus, different restoration objectives should be assigned to distinct sites within a restoration landscape. However, when dealing with multiple functions, trade-offs are likely to arise, whereby the pursuit of one function may prevent the achievement of another. In the restauraR package, the matrix of multifunctionality can be calculated using simple logical tests on each available functional parameter. Thus, the multifunctionality of each restoration site is defined as the number of functions above (or below) a given threshold. The alpha multifunctionality index is defined as the number of individual tests that are true (i.e., the count of restored functions) for a given community.

The selection of simulated species composition can be performed using the alpha multifunctionality index to maximise multifunctionality across restored sites. The alpha multifunctionality index also allows the selection of a simulated community when no solution satisfies all initial criteria, thus allowing users to select a less restricted solution.

High values of the alpha multifunctionality index can be a preferred criterion for selection, meaning that, for each restoration site, the selection is performed in order to return a solution with the highest number of ecosystem functions recovered. 

\strong{Visualisation}

The results visualisation functions are auxiliary to checking parameters in reference sites, analysing the distribution of functional parameters in simulated communities and showing final solutions of selected species composition. 

The function \code{viewResults} allows the visualisation of basic results using histograms for parameter distributions, scatter plots to show the trade-offs between several parameters, restoration costs, species richness and any calculated functional parameter. This visualisation includes all simulated communities and also allows the inclusion of observed relations in reference sites. 

Reference sites are highlighted as distinct points in scatter plots or as rug lines along the axes in histograms.

The function \code{viewMultifunctionality} allows visualisation of multifunctionality sets, creating a graphical representation of the number of sites where each function has been restored, and the number of sites where combinations of functions were restored.

\strong{Merge functions}

The restauraR package includes functions to merge result sets from different operations. 

The function \code{mergeSimulations} concatenates simulated communities generated under different scenarios. For example, some restoration sites can be generated from empty communities, while in another scenario, the species composition can be generated based on established communities. Thus, both scenarios can be concatenated for subsequent steps, such as parameter calculations. 

The function \code{mergeSelection} concatenates different selection procedures. For example, if restoration sites have diverse desired restoration targets, the selection can be performed in distinct steps and then concatenated.

\strong{Multi-site approach}

Selecting communities based solely on the local community may not be optimal for restoration planning across a landscape. The \code{optimiseSelection} function optimises the selection of simulated communities based on ecological indices calculated across multiple sites. The function generates all unique combinations of one simulated community per restoration site and calculates multi-site indices for each combination. Then, the selection of the simulated solutions can be performed on these multisite indices. 

Two sets of indices are available: those derived from the multifunctionality matrix and those derived from species composition. The indices derived from the multifunctionality matrix include the landscape multifunctionality index, which represents the overall restoration success across the entire landscape; the beta multifunctionality index, which is calculated as Rao's quadratic entropy on the multifunctionality matrix, quantifying functional diversity among sites; and the average number of restored functions per site. The indices derived from species composition are calculates Rao's quadratic entropy, either with or without considering species dissimilarities, and include total diversity across all samples, average Alpha and Beta diversity among-community diversity, and Fst, which can be interpreted as local species identity excess. These species composition indices are calculated using a modified version of the raoD function from the picante package.
}
\examples{
data("cerrado")
head(cerrado$traits)
# Restoration new sites
scenarioA <- simulateCommunities(traits = cerrado$traits,
                         ava = "Available",
                         maxDiver = c("SLA", "Height", "Seed"),
                         constCWM = "BT",
                         rich = c(10, 15),
                         it = 100)
scenarioA
# Restoration existing sites
scenarioB <- simulateCommunities(traits = cerrado$traits, 
                                 restComp = cerrado$restoration, 
                                 ava = "Available", 
                                 maxDiver = c("SLA", "Height", "Seed"), 
                                 constCWM = "BT",
                                 rich = c(10, 15), 
                                 it = 100)
scenarioB
# Merge all scenarios
allScenarios <- mergeSimulations(scenarioA, scenarioB)
allScenarios
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Debastiani, V. J., Carlucci, M. B., & Cianciaruso, M. V. (2026). Boosting 
multifunctionality through adaptive trait-based species addition in ongoing restoration projects. 
Ecological Applications, 36, e70197. https://doi.org/10.1002/eap.70197
}
\seealso{
\code{\link{checkReference}}, \code{\link{computeParameters}}, \code{\link{selectCommunities}}, 
\code{\link{extractResults}}, \code{\link{viewResults}}, \code{\link{optimiseSelection}}
}
\author{
See \code{\link{restauraR-package}}.
}
\keyword{MainFunction}
