% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeResbiota.R
\encoding{UTF-8}
\name{themeResbiota}
\alias{themeResbiota}
\title{Internal function. A customized theme to use in ggplot2 objects.}
\usage{
themeResbiota(baseSize = 10, legendPosition = "right")
}
\arguments{
\item{baseSize}{Baseline text size (default baseSize = 10).}

\item{legendPosition}{The position of legends (default legendPosition = "left").}
}
\value{
Theme to ggplot objects.
}
\description{
Internal function. A customized theme to use in ggplot2 objects.
}
\seealso{
\code{\link{simulateCommunities}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{Auxiliary}
