% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewResults.R
\encoding{UTF-8}
\name{viewResults}
\alias{viewResults}
\title{function to visualize results}
\usage{
viewResults(
  x,
  y,
  sim,
  cols = c("grey", "black", "#A6CEE3", "#1F78B4", "#B2DF8A"),
  xlab,
  ylab,
  legend,
  xlim = range(sim$sim_communities$parameters[, x]),
  ylim = range(sim$sim_communities$parameters[, y]),
  hide_notsel = FALSE,
  hide_una = FALSE,
  hide_ref = FALSE,
  hide_sel = FALSE
)
}
\arguments{
\item{x}{name of the variable in x axis. Must match a column in sim and sel_sim}

\item{y}{name of the variable in the y axis. Must match a column in sim and sel_sim}

\item{sim}{list with results of function comSelection.}

\item{cols}{colors of simulated communities, in this order: communities with at least one unavailable species; communities with only available species; communities with at least one unavailable species that satisfy restoration criteria; communities with only available species that satisfy restoration criteria.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}

\item{hide_una:}{hide simulated communities with unavailable species?}

\item{hide_notsel:}{hide simulated communities not selected?}

\item{hide_ref:}{hide reference sites?}
}
\description{
Visualize parameters of simulated and selected communities and reference sites
}
