% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewResults.R, R/viewMultifunctionality.R
\encoding{UTF-8}
\name{viewResults}
\alias{viewResults}
\alias{viewMultifunctionality}
\title{Visualise simulation results}
\usage{
viewResults(x, xvar, yvar = NULL, hideref = FALSE)

viewMultifunctionality(x, hideref = FALSE, ...)
}
\arguments{
\item{x}{An object of class "simRest" or "simRestSelect" to visualise results.}

\item{xvar}{Character vector specifying the variable name for the x-axis in scatter plots.}

\item{yvar}{Character vector specifying the variable name for the y-axis in scatter plots (default yvar = NULL).}

\item{hideref}{Logical indicating whether to exclude reference sites from visualisation (default hideref = FALSE).}

\item{...}{Arguments passed to the upset function of the package ComplexUpset.}
}
\value{
A ggplot plot object that can be further customised using standard ggplot2 syntax.
}
\description{
Generate graphical representations of simulated, selected communities and reference sites, including parameter distributions, trade-off analyses, and multifunctionality assessments.
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\seealso{
\code{\link{simulateCommunities}},\code{\link{computeParameters}}, \code{\link{selectCommunities}}
\code{\link{extractResults}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{MainFunction}
