% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewResults.R, R/viewMultifunctionality.R
\encoding{UTF-8}
\name{viewResults}
\alias{viewResults}
\alias{viewMultifunctionality}
\title{Visualise simulation results}
\usage{
viewResults(x, xvar, yvar = NULL, showReference = TRUE, showMultisite = FALSE)

viewMultifunctionality(x, showReference = FALSE, ...)
}
\arguments{
\item{x}{An object of class "simRest" or "simRestSelect" to visualise results.}

\item{xvar}{Character vector specifying the variable name for the histogram or the x-axis in scatter plots.}

\item{yvar}{Character vector specifying the variable name for the y-axis in scatter plots (default yvar = NULL).}

\item{showReference}{Logical indicating whether to include reference sites from visualisation (default showReference = TRUE).}

\item{showMultisite}{Logical indicating whether to plot results from the multisite analysis; otherwise, the plot shows the basic results (default showMultisite = FALSE).}

\item{...}{Arguments passed to the upset function of the package ComplexUpset.}
}
\value{
A ggplot plot object that can be further customised using standard ggplot2 syntax.
}
\description{
Generate graphical representations of simulated and selected communities, including histograms for parameter distributions, scatter plots for trade-off analyses, and upset plots for multifunctionality results. 
Reference sites are highlighted as distinct points in scatter plots or as rug lines along the axes in histograms.
}
\examples{
data("cerrado")
head(cerrado$traits)
# Simulation
scenario <- simulateCommunities(traits = cerrado$traits,
                                ava = "Available",
                                maxDiver = c("SLA", "Height", "Seed"),
                                constCWM = "BT",
                                rich = c(10, 15),
                                it = 100)
scenario
# Compute functional parameters
scenario <- computeParameters(x = scenario,
                              traits = cerrado$traits,
                              ava = "Available",
                              cwm = "BT",
                              rao = c("SLA", "Height", "Seed"),
                              cost = "Cost",
                              dens = "Density",
                              dissimilarity = c("SLA", "Height", "Seed"),
                              reference = cerrado$reference)
scenario
# Standardise parameters
scenario <- standardiseParameters(x = scenario,
                                  parameters = "dissimilarity",
                                  method = "max")
scenario
# Compute multifunctionality
scenario <- computeMultifunctionality(x = scenario,
                                      tests = c("CWM_BT > 5.9",
                                                "rao > 0.2"))
scenario
# View results - histogram
viewResults(scenario, xvar = "CWM_BT")
# View results - scatter plot
viewResults(scenario, xvar = "CWM_BT", yvar = "rao")
# View multifunctionality - upset plot
viewMultifunctionality(scenario)
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\seealso{
\code{\link{simulateCommunities}}, \code{\link{computeParameters}}, \code{\link{selectCommunities}}, \code{\link{extractResults}}
}
\author{
See \code{\link{resbiota-package}}.
}
\keyword{MainFunction}
