% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewResults.R, R/viewMultifunctionality.R
\encoding{UTF-8}
\name{viewResults}
\alias{viewResults}
\alias{viewMultifunctionality}
\title{Visualize results}
\usage{
viewResults(x, xvar, yvar, hideref = FALSE)

viewMultifunctionality(x, ...)
}
\arguments{
\item{x}{A object of class "simRest" or "simRestSelect" to visualize results}

\item{xvar}{Name of the variable (parameter) in x axis.}

\item{yvar}{Name of the variable (parameter) in y axis.}

\item{hideref}{Logical argument to specify if hide reference sites}

\item{...}{Arguments passed to \code{\link{ComplexUpset::upset}} function.}
}
\value{
A ggplot plot object.
}
\description{
Visualize parameters of simulated and selected communities and reference sites
}
